requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL adapter
python-jose[cryptography]==3.3.0  # For JWT tokens
passlib[bcrypt]==1.7.4  # For password hashing
python-multipart==0.0.9  # For form data
pytest==8.0.2  # For testing
httpx==0.27.0  # For async HTTP requests in tests
black==24.2.0  # For code formatting
flake8==7.0.0  # For linting
mypy==1.8.0  # For type checking
cryptography==42.0.5  # For AES encryption
email-validator==2.1.0.post1  # For email validation 